+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/tool/editor/PsiJavaUtils.kt
@@ -9,6 +9,8 @@ import com.intellij.psi.PsiIdentifier
 import com.intellij.psi.PsiJavaToken
 import com.intellij.psi.util.elementType
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.common.EditorUtils.getSelectedText
+import org.jetbrains.kotlin.idea.editor.fixers.end
+import org.jetbrains.kotlin.idea.editor.fixers.start
 
 /**
  * Some code parts of this class are only available of the optional dependency
@@ -27,13 +29,21 @@ internal object PsiJavaUtils {
     }
 
     val psiElement = psiFile.findElementAt(editor.caretModel.offset) ?: return null
-    return getTextIfStringValueOrIdentifier(psiElement)?.let { it to psiElement.textRange }
+    return getTextIfStringValueOrIdentifier(psiElement)
   }
 
-  fun getTextIfStringValueOrIdentifier(psiElement: PsiElement): String? {
-    val isStringLiteral = psiElement is PsiJavaToken && psiElement.elementType == JavaTokenType.STRING_LITERAL
-    return if (isStringLiteral || psiElement is PsiIdentifier) {
-      psiElement.text
+  fun getTextIfStringValueOrIdentifier(psiElement: PsiElement): Pair<String, TextRange>? {
+    return if (psiElement is PsiJavaToken && psiElement.elementType == JavaTokenType.STRING_LITERAL) {
+      // Remove the enclosing quotations
+      val newStart = psiElement.textRange.start + 1
+      val newEnd = psiElement.textRange.end - 1
+      if (newStart > newEnd) {
+        return null
+      }
+      psiElement.text.substring(1, psiElement.text.length - 1) to TextRange(newStart, newEnd)
+    }
+    else if (psiElement is PsiIdentifier) {
+      psiElement.text to psiElement.textRange
     }
     else {
       null